addedd uninstall confirmation to the uninstall script

+++ uninstall.sh
@@ -1,9 +1,9 @@
-#!/bin/bash
+#!/bin/sh
+#
+#
+#
+#
 
-### Setup Path and find TTY
-
-  PATH=/bin:/sbin:/usr/bin:/usr/sbin
-  TTY="/dev/$( ps -p$$ -o tty | tail -1 | awk '{print$1}' )"
 
 ### Check if HearthLog or any HearthLog components are installed
 
@@ -12,18 +12,25 @@
     exit 1
   fi
 
+### Setup Path and find TTY
 
-echo "Uninstalling HearthLog..."
+  PATH=/bin:/sbin:/usr/bin:/usr/sbin
+  TTY="/dev/$( ps -p$$ -o tty | tail -1 | awk '{print$1}' )"
+
+### Verify that the user wishes to uninstall
+  read -p "Uninstall HearthLog [y/n]? " ANSWER < $TTY
+  [[ $ANSWER == "y" ]] || exit 1
+  echo "Uninstalling HearthLog..."
 
 ### remove LaunchDaemon plist
-sudo rm /Library/LaunchDaemons/org.hearthlog.bpf.chown.plist
+  sudo rm /Library/LaunchDaemons/org.hearthlog.bpf.chown.plist
 
 ### remove bpf chown script
-sudo rm /Library/Scripts/hearthlogBpfChown.bash
+  sudo rm /Library/Scripts/hearthlogBpfChown.bash
 
 ### remove the .app in from /Applications
-sudo rm -R /Applications/HearthLog.app
+  sudo rm -R /Applications/HearthLog.app
 
-echo "HearthLog has been uninstalled"
+  echo "HearthLog has been uninstalled"
 
-exit 0
+  exit 0
